Collapse Checkout session address type
Committed-By-Agent: codex
